👄 Simplify Response::LineItemProduct
When parsing returned `LineItemProduct` values from Quotation/Invoice/DistributeTax, we get one of 2 types of objects back from `Savon`:
* Typically: a `Nori::StringWithAttributes`, when a product **has** a `product code` specified.
* Less common: a `Hash`, when a product **does not have** a  `product code` specified.

Also, when we are in the fallback scenario, we are dealing with a varied input format based on our `Payload` rather than the `Savon` response.

Because the construction of these `LineItemProduct` objects is so varied, we move the responsibility for creating them properly into the objects generating the `Response`, and turn `LineItemProduct` into a simple value store.