VertexClient 1.0.0
This pull request bumps VertexClient to version 1.0.0, removing the support for all Ruby languages that are marked "end of life".
This allows us to focus on actively maintained versions only and not introduce potential security concerns.

Projects still running on versions older than 2.5 are recommended to upgrade their versions as soon as possible.